final merge (ai repaired, was placed wrong)

+++ src/game_logic/World.h
@@ -48,7 +48,7 @@ private:
         float _generation_time;
         float _time_limit;
         int _populationCars;
-        int _mutationRate;
+        float _mutationRate;
 
 public:
         World(std::shared_ptr<IEntityModelCreator> entity_model_creator, float x_min, float x_max, float y_min,
